+++ Package.swift
@@ -14,7 +14,7 @@ let package = Package(
     dependencies: [
         .package(
             url: "https://github.com/Miguelosaurus/InjectionNext.git",
-            revision: "547f900b7f784eb1ca29415f0030bed3e847e3e1"
+            revision: "9855a40"
         ),
     ],
     targets: [